Last of the mohicans

+++ GeneticAlgorithm/src/main/java/com/nure/genetic/abstraction/FitnessFunction.java
@@ -2,5 +2,5 @@ package com.nure.genetic.abstraction;
 
 @FunctionalInterface
 public interface FitnessFunction {
-    double getValue(double... xs);
+    float getValue(float... xs);
 }
