+++ GeneticAlgorithm/src/main/java/com/nure/genetic/implementation/MyFitnessFunction.java
@@ -4,7 +4,7 @@ import com.nure.genetic.abstraction.FitnessFunction;
 
 public class MyFitnessFunction implements FitnessFunction {
     @Override
-    public double getValue(double... xs) {
+    public float getValue(float... xs) {
         return  - xs[0] * xs[0] - 2 * xs[0] + 3;
     }
 }
